test: 인수 테스트 step 분리
